moved prompt to util module
