+++ util.py
@@ -1,6 +1,14 @@
 import sys
 import functools
 
+def prompt(q,opts=['y','n']) :
+    resp = "nopenopenope"
+    while resp not in opts :
+        print q+" [%s]"%''.join(opts)
+        resp = raw_input().lower()
+    return resp
+
+
 _ascii = {  'blue': '\033[0;34m'
            , 'yellow': '\033[0;33m'
            , 'green': '\033[0;32m'
